# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZATCA Saudi e-invoicing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CompanyInvoiceType =
  (typeof CompanyInvoiceType)[keyof typeof CompanyInvoiceType];

export const CompanyInvoiceType = {
  standard: "standard",
  simplified: "simplified",
  both: "both",
} as const;

export interface Company {
  id: number;
  /** Company name in Arabic */
  nameAr: string;
  /** Company name in English */
  nameEn?: string;
  /** VAT registration number (15 digits) */
  vatNumber: string;
  /** Commercial registration number */
  crNumber: string;
  city: string;
  district?: string;
  street: string;
  buildingNumber: string;
  postalCode: string;
  additionalNumber?: string;
  country: string;
  industryName?: string;
  invoiceType: CompanyInvoiceType;
  isSandbox: boolean;
  /** Device serial number */
  serialNumber?: string;
  deviceSerial1?: string;
  deviceSerial2?: string;
  deviceSerial3?: string;
  /** ZATCA Cryptographic Stamp Identifier */
  zatcaCsid?: string;
  /** ZATCA Production CSID */
  zatcaPcsid?: string;
  createdAt?: string;
  updatedAt?: string;
}

export type CreateCompanyBodyInvoiceType =
  (typeof CreateCompanyBodyInvoiceType)[keyof typeof CreateCompanyBodyInvoiceType];

export const CreateCompanyBodyInvoiceType = {
  standard: "standard",
  simplified: "simplified",
  both: "both",
} as const;

export interface CreateCompanyBody {
  nameAr: string;
  nameEn?: string;
  vatNumber: string;
  crNumber: string;
  city: string;
  district?: string;
  street: string;
  buildingNumber: string;
  postalCode: string;
  additionalNumber?: string;
  country?: string;
  industryName?: string;
  invoiceType?: CreateCompanyBodyInvoiceType;
  isSandbox?: boolean;
  serialNumber?: string;
  deviceSerial1?: string;
  deviceSerial2?: string;
  deviceSerial3?: string;
  zatcaCsid?: string;
  zatcaPcsid?: string;
}

export interface Customer {
  id: number;
  companyId: number;
  nameAr: string;
  nameEn?: string;
  vatNumber?: string;
  crNumber?: string;
  email?: string;
  phone?: string;
  city?: string;
  district?: string;
  street?: string;
  buildingNumber?: string;
  postalCode?: string;
  country?: string;
  createdAt?: string;
}

export interface CreateCustomerBody {
  companyId: number;
  nameAr: string;
  nameEn?: string;
  vatNumber?: string;
  crNumber?: string;
  email?: string;
  phone?: string;
  city?: string;
  district?: string;
  street?: string;
  buildingNumber?: string;
  postalCode?: string;
  country?: string;
}

export interface InvoiceLineItem {
  id: number;
  invoiceId?: number;
  description: string;
  quantity: number;
  unitPrice: number;
  discountAmount?: number;
  /** VAT rate percentage (e.g. 15 for 15%) */
  vatRate: number;
  vatAmount?: number;
  subtotal?: number;
  total?: number;
}

export interface CreateInvoiceLineItem {
  description: string;
  quantity: number;
  unitPrice: number;
  discountAmount?: number;
  vatRate?: number;
}

export type InvoiceInvoiceType =
  (typeof InvoiceInvoiceType)[keyof typeof InvoiceInvoiceType];

export const InvoiceInvoiceType = {
  standard: "standard",
  simplified: "simplified",
} as const;

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  draft: "draft",
  issued: "issued",
  cancelled: "cancelled",
} as const;

export type InvoiceZatcaStatus =
  (typeof InvoiceZatcaStatus)[keyof typeof InvoiceZatcaStatus];

export const InvoiceZatcaStatus = {
  pending: "pending",
  reported: "reported",
  cleared: "cleared",
  rejected: "rejected",
} as const;

export interface Invoice {
  id: number;
  companyId: number;
  customerId?: number;
  invoiceNumber: string;
  invoiceType: InvoiceInvoiceType;
  status: InvoiceStatus;
  issueDate: string;
  supplyDate?: string;
  dueDate?: string;
  currency: string;
  subtotal: number;
  discountTotal?: number;
  vatTotal: number;
  grandTotal: number;
  notes?: string;
  /** Base64 encoded QR code */
  qrCode?: string;
  /** Cryptographic hash for ZATCA */
  invoiceHash?: string;
  zatcaStatus?: InvoiceZatcaStatus;
  zatcaResponseCode?: string;
  lineItems?: InvoiceLineItem[];
  company?: Company;
  customer?: Customer;
  createdAt?: string;
  updatedAt?: string;
}

export type CreateInvoiceBodyInvoiceType =
  (typeof CreateInvoiceBodyInvoiceType)[keyof typeof CreateInvoiceBodyInvoiceType];

export const CreateInvoiceBodyInvoiceType = {
  standard: "standard",
  simplified: "simplified",
} as const;

export interface CreateInvoiceBody {
  companyId: number;
  customerId?: number;
  invoiceType: CreateInvoiceBodyInvoiceType;
  issueDate: string;
  supplyDate?: string;
  dueDate?: string;
  currency?: string;
  notes?: string;
  lineItems: CreateInvoiceLineItem[];
}

export interface DashboardSummary {
  totalInvoices: number;
  draftCount: number;
  issuedCount: number;
  cancelledCount: number;
  totalVat: number;
  totalRevenue: number;
  standardCount: number;
  simplifiedCount: number;
  thisMonthRevenue: number;
  thisMonthVat: number;
}

export interface MonthlyStat {
  month: string;
  year: number;
  invoiceCount: number;
  revenue: number;
  vatAmount: number;
}

export type ListCustomersParams = {
  companyId?: number;
};

export type ListInvoicesParams = {
  companyId?: number;
  status?: ListInvoicesStatus;
  invoiceType?: ListInvoicesInvoiceType;
};

export type ListInvoicesStatus =
  (typeof ListInvoicesStatus)[keyof typeof ListInvoicesStatus];

export const ListInvoicesStatus = {
  draft: "draft",
  issued: "issued",
  cancelled: "cancelled",
} as const;

export type ListInvoicesInvoiceType =
  (typeof ListInvoicesInvoiceType)[keyof typeof ListInvoicesInvoiceType];

export const ListInvoicesInvoiceType = {
  standard: "standard",
  simplified: "simplified",
} as const;

export type GetDashboardSummaryParams = {
  companyId?: number;
};

export type GetRecentInvoicesParams = {
  companyId?: number;
  limit?: number;
};

export type GetMonthlyStatsParams = {
  companyId?: number;
  year?: number;
};